move big end detection to its own file
